Rename file 'cliente.py' to 'client.py'
